confirmation

+++ front-end/src/app/movie/movie.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MoviesService } from '../movies.service';
 import { Movie } from '../movie';
+import swal from 'sweetalert2';
 
 @Component({
   selector: 'app-movie',
@@ -21,11 +22,21 @@ export class MovieComponent implements OnInit {
   }
 
   public deleteMovie() {
-    if (this.movie$.id)  {
-      this.movieService.deleteMovie(this.movie$.id)
-          .subscribe((message: any) => {
-            this.router.navigate(['movies']);
-        });
+    if (this.movie$.id) {
+      swal.fire({
+        titleText: 'Are you sure?',
+        text: 'Once deleted, you cannot recover this movie.',
+        icon: 'warning',
+        showCancelButton: true
+      }).then((willDelete: any) => {
+        console.log(willDelete);
+        if (willDelete.value) {
+          this.movieService.deleteMovie(this.movie$.id)
+            .subscribe((message: any) => {
+              this.router.navigate(['movies']);
+            });
+        }
+      });
     }
   }
 }
